refactor(employee): use IAppDbContext instead of IUnitOfWork in employee logic

+++ LibraryMS.Application/Features/Employee/Commands/Delete/DeleteEmployeeCommandHandler.cs
@@ -1,31 +1,46 @@
+using LibraryMS.Application.Common.Extensions;
+
 namespace LibraryMS.Application.Features.Employee.Commands.Delete;
 
-public sealed class DeleteEmployeeCommandHandler(IUnitOfWork unitOfWork, IIdentityUser identityUser) : IRequestHandler<DeleteEmployeeCommand, Result>
+public sealed class DeleteEmployeeCommandHandler(IAppDbContext context, IIdentityUser identityUser) : IRequestHandler<DeleteEmployeeCommand, Result>
 {
     public async Task<Result> Handle(DeleteEmployeeCommand request, CancellationToken cancellationToken)
     {
-        await using var transaction = await unitOfWork.BeginTransactionAsync();
+        await using var transaction = await context.BeginTransactionAsync(cancellationToken);
 
         try
         {
             // Delete Employee
-            var employee = await unitOfWork.Employees.GetEmployeeByUserIdAsync(request.UserId);
+            var employee = await context.Employees
+            .SingleOrDefaultAsync(e => e.Id == request.EmployeeId, cancellationToken);
             if (employee is null)
             {
                 return Result.Failure("Employee not found.");
             }
-            unitOfWork.Employees.Delete(employee);
 
-            // Delete User
+            if (employee.IsDeleted)
+            {
+                return Result.Failure("Employee is already deleted.");
+            }
 
-            var userResult = await identityUser.DeleteUserAsync(request.UserId);
+            var userResult = await identityUser.ValidateUserPersonMatchAsync(request.UserId, employee.PersonId);
             if (userResult.IsFailure)
             {
-                await transaction.RollbackAsync(cancellationToken);
                 return Result.Failure(userResult.Error);
             }
 
-            await unitOfWork.SaveChangesAsync(cancellationToken);
+            context.Employees.Remove(employee);
+
+            // Delete User
+
+            var deleteUserResult = await identityUser.DeleteUserAsync(request.UserId);
+            if (deleteUserResult.IsFailure)
+            {
+                await transaction.RollbackAsync(cancellationToken);
+                return Result.Failure(deleteUserResult.Error);
+            }
+
+            await context.SaveChangesAsync(cancellationToken);
             await transaction.CommitAsync(cancellationToken);
             return Result.Success;
         }
